add: create github action yaml programatically
