+++ utils/scripts/aws-lambda-okta-webhook/outputs.tf
@@ -1,0 +1,5 @@
+output "idt-okta-webhook-url" {
+  value       = "module.idt-okta-webhook-url.function_url"
+  description = "The URL of the deployed Lambda"
+}
+
