+++ utils/scripts/aws-lambda-okta-webhook/src/index.ts
@@ -1,0 +1,6 @@
+import { OktaGroupIntegration, OktaUserIntegration } from '@indent/integration-okta'
+import { getLambdaHandler } from '@indent/runtime-aws-lambda'
+
+export const handle = getLambdaHandler({
+  integrations: [new OktaGroupIntegration(), new OktaUserIntegration()],
+})
