+++ utils/scripts/js/format-types.ts
