+++ utils/scripts/package.json
@@ -14,6 +14,7 @@
     "terraform-generator": "^5.2.0",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.4",
+    "yaml": "^2.1.1",
     "yarn": "^1.22.18"
   },
   "devDependencies": {
